Direction and Percent Change

+++ src/MarketData.TdAmeritrade/Candle.cs
@@ -4,10 +4,10 @@ namespace MarketData.TdAmeritrade
 {
     public class Candle
     {
-        public float Open { get; set; }
-        public float High { get; set; }
-        public float Low { get; set; }
-        public float Close { get;set; }
+        public decimal Open { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Close { get;set; }
         public int Volume { get; set; }
         public long DateTime { get; set; }
     }
